% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_cata_des.R
\name{fun_cata_des}
\alias{fun_cata_des}
\title{Title 分类变量统计描述与推断}
\usage{
fun_cata_des(
  df,
  group,
  col,
  index,
  name,
  show_stat = T,
  show_method = T,
  print = F,
  digit_p = 4,
  digit_s = 3,
  digit_pct = 2,
  show_pct = T,
  ...
)
}
\arguments{
\item{df}{被分析的dataframe}

\item{group}{哪一列作为输出的列（如试验组、对照组)}

\item{col}{哪一列作为交叉表的行}

\item{index}{哪几列作为交叉表的标识}

\item{name}{为项目指标分类设置内容}

\item{sum_row}{是否计算行合计}

\item{sum_col}{是否计算列合计}

\item{col_sum_pct}{是否计算列合计的百分比}

\item{pct}{是否计算百分比}

\item{stat}{是否进行统计检验}

\item{denominator}{传入一个向量用做计算列的分母}

\item{show_na}{是否在列合计展示缺失值}
}
\value{
a dataframe
}
\description{
Title 分类变量统计描述与推断
}
\examples{
mpg \%>\%
  select(where(is.character)) \%>\%
  mutate(GROUP = rep(c("试验组", "对照组"), nrow(mpg)/2)) \%>\%
  select(model, drv, GROUP) \%>\%
 fun_cata_des("GROUP",col = "model", index = "drv",name = "")
}
