% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_com_site.R
\name{fun_com_site}
\alias{fun_com_site}
\title{Title 用于合并中心}
\usage{
fun_com_site(df, ls_com)
}
\arguments{
\item{df}{必须包含中心列和数据列的数据框}

\item{ls_com}{代表合并中心的列表}
}
\value{
a data.frame
}
\description{
Title 用于合并中心
}
\examples{
-
}
