% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_cross.R
\name{fun_cross}
\alias{fun_cross}
\title{Title}
\usage{
fun_cross(
  df,
  group,
  col,
  sum_row = T,
  sum_col = T,
  show_pct = T,
  col_sum_pct = F,
  denominator = NULL,
  index = NULL,
  na.rm = T,
  show_na = T,
  digit_pct = 2
)
}
\arguments{
\item{df}{The dataframe you want to  a crosstable}

\item{group}{Variables as columns}

\item{col}{作为行的变量}

\item{sum_row}{是否计算行合计}

\item{sum_col}{是否计算列合计}

\item{col_sum_pct}{是否计算列合计的百分比}

\item{denominator}{传入数值向量作为计算百分比的分母}

\item{index}{哪几列作为交叉表的标识}

\item{na.rm}{是否移除缺失值}

\item{show_na}{如果为TRUE,列合计的括号里将会展示缺失值的个数}

\item{pct}{是否计算百分比}
}
\value{
a dataframe
}
\description{
Title
}
\examples{
df = data.frame(group = rep(c("a", "b"), 6), col = rep(c("c", "a", "f"),4 ))
fun_cross(df, group = "group", col = "col", sum_col = T,sum_row = T,  show_na = F, show_pct = T, digit_pct = 3)
}
