% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_load_file.R
\name{fun_load_file}
\alias{fun_load_file}
\title{Title 批量读入文件，支持.sas7bdat、.xlsx、.csv}
\usage{
fun_load_file(path, suffix, filEencoding = NULL)
}
\arguments{
\item{path}{要批量导入的路径}

\item{suffix}{文件后缀}

\item{filEencoding}{编码方式， 仅.csv文件}
}
\value{
\itemize{
\item 
}
}
\description{
Title 批量读入文件，支持.sas7bdat、.xlsx、.csv
}
\examples{
-
}
