% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_na.R
\name{fun_NA}
\alias{fun_NA}
\title{Title 制表,填缺失值}
\usage{
fun_NA(data, vec)
}
\arguments{
\item{data}{数据框}

\item{vec}{字符向量，向量的内容为要填补缺失值得列名}
}
\value{
a dataframe
}
\description{
Title 制表,填缺失值
}
\examples{
iris \%>\% fun_NA("Species")
}
