% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_non_inferiority.R
\name{fun_non_inferiority}
\alias{fun_non_inferiority}
\title{Title}
\usage{
fun_non_inferiority(x1, x2, deta)
}
\arguments{
\item{x1}{试验组数据的数值向量}

\item{x2}{肚子好租数据的数值向量}

\item{deta}{非劣效界值}
}
\value{
包含结果的列表
}
\description{
Title
}
\examples{
-
}
