% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_rank_des.R
\name{fun_rank_des}
\alias{fun_rank_des}
\title{Title}
\usage{
fun_rank_des(
  df,
  group,
  col,
  index,
  name,
  sum_row = T,
  sum_col = F,
  pct = T,
  stat = T,
  denominator = NULL,
  show_na = F
)
}
\arguments{
\item{df}{\itemize{
\item 
}}

\item{group}{\itemize{
\item 
}}

\item{col}{\itemize{
\item 
}}

\item{index}{\itemize{
\item 
}}

\item{name}{\itemize{
\item 
}}

\item{sum_row}{\itemize{
\item 
}}

\item{sum_col}{\itemize{
\item 
}}

\item{pct}{\itemize{
\item 
}}

\item{stat}{\itemize{
\item 
}}

\item{denominator}{\itemize{
\item 
}}

\item{show_na}{\itemize{
\item 
}}
}
\value{
a dataframe
}
\description{
Title
}
\examples{
-
}
