% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_save_docx.R
\name{fun_save_docx}
\alias{fun_save_docx}
\title{Title}
\usage{
fun_save_docx(df, path)
}
\arguments{
\item{df}{要输出的data.frame}

\item{path}{输出的路径,以,docx结尾}
}
\value{
none
}
\description{
Title
}
\examples{
none
}
