% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_select.R
\name{fun_select}
\alias{fun_select}
\title{Title}
\usage{
fun_select(
  df,
  var,
  new_names = NULL,
  names_posit = NULL,
  df_group = NULL,
  join_type = "right"
)
}
\arguments{
\item{df}{原始数据框， 其中必须有SUBJID列}

\item{var}{选择的列名，SUBJID必被选}

\item{new_names}{为列修改列名}

\item{names_posit}{按位置修改列名}

\item{join_type}{连接类型}

\item{join_df}{连接的数据框， 按照SUBJID连接}
}
\value{
a dataframe
}
\description{
Title
}
\examples{
df = mpg \%>\% mutate(SUBJID = 1:nrow(mpg))
join_df = data.frame(SUBJID = 99:nrow(mpg), 治疗分组 = c("试验组", "对照组"))
df \%>\% fun_select(c(SUBJID, hwy),new_names = "hwy_new", df_group = join_df , join_type = "full")
}
