% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_stat.R
\name{fun_stat_col}
\alias{fun_stat_col}
\title{Title
对各列进行统计描述和检验的函数}
\usage{
fun_stat_col(
  df_all,
  group,
  index = NULL,
  show_stat = T,
  show_method = T,
  show_conf = F,
  show_pct = T,
  show_na = T,
  sum_col = T,
  sum_row = T,
  col_sum_pct = F,
  denominator = NULL,
  digit_s = 3,
  digit_p = 4,
  digit_des = 2,
  digit_pct = 2,
  print = F
)
}
\arguments{
\item{df_all}{宽数据形式的data.frame}

\item{group}{指定分组列，一般为"试验药物"列}

\item{index}{指定一列或者哪几列作为输出的标识，标识列不参与统计描述与检验}

\item{show_stat}{是否进行假设检验}

\item{show_method}{是否展示假设检验采用的方法}

\item{show_conf}{是否展示置信区间，仅针对连续型变量}

\item{show_pct}{对于分类变量，交叉表中是否添加百分比}

\item{show_na}{是否展示各组缺失值数量}

\item{sum_col}{对于分类变量, 交叉表是否展示列合计}

\item{sum_row}{对于分类变量, 交叉表是否展示行合计}

\item{col_sum_pct}{对于分类变量, 交叉表的列合计是否展示百分比}

\item{denominator}{对于分类变量, 通过传入一个向量作为分母计算交叉表百分比}

\item{digit_s}{检验统计量保留的喜小数位}

\item{digit_p}{p值保留的小数位}

\item{digit_des}{对于连续型变量，统计描述部分保留的小数位}

\item{digit_pct}{对于分类变量， 交叉表的百分比保留的小数位}

\item{print}{在进行每个变量的统计分析时，是否在命令行打印使用的检验方法}
}
\value{
a dataframe
}
\description{
Title
对各列进行统计描述和检验的函数
}
\examples{
mpg \%>\% mutate(GROUP = rep(c("试验组", "对照组"), nrow(mpg)/2)) \%>\%
select(where(is.character)) \%>\%
  fun_stat_col(group = "GROUP", print = T)
}
