% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_stat_col_mul.R
\name{fun_stat_col_mul}
\alias{fun_stat_col_mul}
\title{Title}
\usage{
fun_stat_col_mul(df, col_mul, ..., col_mul_vet = NULL)
}
\arguments{
\item{df}{长数据}

\item{col_mul}{长数据的name列}

\item{...}{}

\item{col_mul_vet}{需要分析name列中的哪些元素}
}
\value{
a dataframe
}
\description{
Title
}
\examples{
mpg \%>\% mutate(GROUP = rep(c("试验组", "对照组"), nrow(mpg)/2)) \%>\%
  fun_stat_col_mul(df= .,group = "GROUP", col_mul = "drv")
}
