% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round0.R
\name{round0}
\alias{round0}
\title{Title}
\usage{
round0(vct, num = 2)
}
\arguments{
\item{vct}{A numeric vector.}

\item{num}{A numeric vector of length one}
}
\value{
A character vector
}
\description{
Title
}
\examples{
(x = c(1.999, 1.005, 1.134))
round0(x, 2)
}
