% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_wilcoxon_2s.R
\name{wilcoxon_2s}
\alias{wilcoxon_2s}
\title{Title}
\usage{
wilcoxon_2s(
  df,
  alternative = "two.sided",
  correct = TRUE,
  mu = 0,
  digits.rank = Inf
)
}
\arguments{
\item{df}{该函数不会被导出到namespace}

\item{alternative}{NONE}

\item{correct}{NONE}

\item{mu}{NONE}

\item{digits.rank}{NONE}
}
\value{
NONE
}
\description{
Title
}
\examples{
NONE
}
